Move DirState and FPurgeState to a separate source file.
