+++ src/XrdPfc/XrdPfcPurge.cc
@@ -1,175 +1,26 @@
 #include "XrdPfc.hh"
+#include "XrdPfcDirState.hh"
+#include "XrdPfcFPurgeState.hh"
 #include "XrdPfcTrace.hh"
 
 #include <fcntl.h>
 #include <sys/time.h>
 
 #include "XrdOuc/XrdOucEnv.hh"
+#include "XrdOuc/XrdOucUtils.hh"
 #include "XrdOss/XrdOssAt.hh"
 #include "XrdSys/XrdSysTrace.hh"
 
 using namespace XrdPfc;
-
 namespace XrdPfc
 {
 
+/*
 XrdSysTrace* GetTrace()
 {
    // needed for logging macros
    return Cache::GetInstance().GetTrace();
-}
-
-// Temporary, extensive purge tracing
-// #define TRACE_PURGE(x) TRACE(Debug, x)
-// #define TRACE_PURGE(x) std::cout << "PURGE " << x << "\n"
-#define TRACE_PURGE(x)
-
-//==============================================================================
-// DirState
-//==============================================================================
-
-class DirState
-{
-   DirState    *m_parent;
-
-   Stats        m_stats;        // access stats from client reads in this directory (and subdirs)
-
-   long long    m_usage;        // collected / measured during purge traversal
-   long long    m_usage_extra;  // collected from write events in this directory and subdirs
-   long long    m_usage_purged; // amount of data purged from this directory (and subdirectories for leaf nodes)
-
-   // begin purge traversal usage \_ so we can have a good estimate of what came in during the traversal
-   // end purge traversal usage   /  (should be small, presumably)
-
-   // quota info, enabled?
-
-   int          m_depth;
-   int          m_max_depth;    // XXXX Do we need this? Should it be passed in to find functions?
-   bool         m_stat_report;  // not used yet - storing of stats requested
-
-   typedef std::map<std::string, DirState> DsMap_t;
-   typedef DsMap_t::iterator               DsMap_i;
-
-   DsMap_t      m_subdirs;
-
-   void init()
-   {
-      m_usage = 0;
-      m_usage_extra  = 0;
-      m_usage_purged = 0;
-   }
-
-   DirState* create_child(const std::string &dir)
-   {
-      std::pair<DsMap_i, bool> ir = m_subdirs.insert(std::make_pair(dir, DirState(this)));
-      return &  ir.first->second;
-   }
-
-   DirState* find_path_tok(PathTokenizer &pt, int pos, bool create_subdirs)
-   {
-      if (pos == pt.get_n_dirs()) return this;
-
-      DsMap_i i = m_subdirs.find(pt.m_dirs[pos]);
-
-      DirState *ds = 0;
-
-      if (i != m_subdirs.end())
-      {
-         ds = & i->second;
-      }
-      if (create_subdirs && m_depth < m_max_depth)
-      {
-         ds = create_child(pt.m_dirs[pos]);
-      }
-      if (ds) return ds->find_path_tok(pt, pos + 1, create_subdirs);
-
-      return 0;
-   }
-
-public:
-
-   DirState(int max_depth) : m_parent(0), m_depth(0), m_max_depth(max_depth)
-   {
-      init();
-   }
-
-   DirState(DirState *parent) : m_parent(parent), m_depth(m_parent->m_depth + 1), m_max_depth(m_parent->m_max_depth)
-   {
-      init();
-   }
-
-   DirState* get_parent()                     { return m_parent; }
-
-   void      set_usage(long long u)           { m_usage = u; m_usage_extra = 0; }
-   void      add_up_stats(const Stats& stats) { m_stats.AddUp(stats); }
-   void      add_usage_purged(long long up)   { m_usage_purged += up; }
-
-   DirState* find_path(const std::string &path, int max_depth, bool parse_as_lfn, bool create_subdirs)
-   {
-      PathTokenizer pt(path, max_depth, parse_as_lfn);
-
-      return find_path_tok(pt, 0, create_subdirs);
-   }
-
-   DirState* find_dir(const std::string &dir, bool create_subdirs)
-   {
-      DsMap_i i = m_subdirs.find(dir);
-
-      if (i != m_subdirs.end())  return & i->second;
-
-      if (create_subdirs && m_depth < m_max_depth)  return create_child(dir);
-
-      return 0;
-   }
-
-   void reset_stats()
-   {
-      m_stats.Reset();
-
-      for (DsMap_i i = m_subdirs.begin(); i != m_subdirs.end(); ++i)
-      {
-         i->second.reset_stats();
-      }
-   }
-
-   void upward_propagate_stats()
-   {
-      for (DsMap_i i = m_subdirs.begin(); i != m_subdirs.end(); ++i)
-      {
-         i->second.upward_propagate_stats();
-
-         m_stats.AddUp(i->second.m_stats);
-      }
-
-      m_usage_extra += m_stats.m_BytesWritten;
-   }
-
-   long long upward_propagate_usage_purged()
-   {
-      for (DsMap_i i = m_subdirs.begin(); i != m_subdirs.end(); ++i)
-      {
-         m_usage_purged += i->second.upward_propagate_usage_purged();
-      }
-      m_usage -= m_usage_purged;
-
-      long long ret = m_usage_purged;
-      m_usage_purged = 0;
-      return ret;
-   }
-
-   void dump_recursively(const char *name)
-   {
-      printf("%*d %s usage=%lld usage_extra=%lld usage_total=%lld num_ios=%d duration=%d b_hit=%lld b_miss=%lld b_byps=%lld b_wrtn=%lld\n",
-             2 + 2*m_depth, m_depth, name, m_usage, m_usage_extra, m_usage + m_usage_extra,
-             m_stats.m_NumIos, m_stats.m_Duration, m_stats.m_BytesHit, m_stats.m_BytesMissed, m_stats.m_BytesBypassed, m_stats.m_BytesWritten);
-
-      for (DsMap_i i = m_subdirs.begin(); i != m_subdirs.end(); ++i)
-      {
-         i->second.dump_recursively(i->first.c_str());
-      }
-   }
-};
-
+}*/
 
 //==============================================================================
 // DataFsState
@@ -215,312 +66,6 @@ public:
 };
 
 
-//==============================================================================
-// FPurgeState
-//==============================================================================
-
-class FPurgeState
-{
-public:
-   struct FS
-   {
-      std::string path;
-      long long   nBytes;
-      time_t      time;
-      DirState   *dirState;
-
-      FS(const std::string &dname, const char *fname, long long n, time_t t, DirState *ds) :
-         path(dname + fname), nBytes(n), time(t), dirState(ds)
-      {}
-   };
-
-   typedef std::multimap<time_t, FS> map_t;
-   typedef map_t::iterator           map_i;
-
-   map_t   m_fmap; // map of files that are purge candidates
-
-   typedef std::list<FS>    list_t;
-   typedef list_t::iterator list_i;
-
-   list_t  m_flist; // list of files to be removed unconditionally
-
-   long long nBytesReq;
-   long long nBytesAccum;
-   long long nBytesTotal;
-   time_t    tMinTimeStamp;
-   time_t    tMinUVKeepTimeStamp;
-
-   // XrdOss   *m_oss;
-   XrdOssAt  m_oss_at;
-
-   // ------------------------------------
-   // Directory handling & stat collection
-   // ------------------------------------
-
-   DirState    *m_dir_state;
-   std::string  m_current_path; // Includes trailing '/'
-   int          m_dir_level;
-   const int    m_max_dir_level_for_stat_collection; // until we honor globs from pfc.dirstats
-
-   std::vector<std::string> m_dir_names_stack;
-   std::vector<long long>   m_dir_usage_stack;
-
-   const char   *m_info_ext;
-   const size_t  m_info_ext_len;
-   XrdSysTrace  *m_trace;
-
-   static const char *m_traceID;
-
-
-   void begin_traversal(DirState *root, const char *root_path = "/")
-   {
-      m_dir_state = root;
-      m_dir_level = 0;
-      m_current_path = std::string(root_path);
-      m_dir_usage_stack.push_back(0);
-
-      TRACE_PURGE("FPurgeState::begin_traversal cur_path '" << m_current_path << "', usage=" << m_dir_usage_stack.back() << ", level=" << m_dir_level);
-   }
-
-   void end_traversal()
-   {
-      TRACE_PURGE("FPurgeState::end_traversal reporting for '" << m_current_path << "', usage=" << m_dir_usage_stack.back() << ", nBytesTotal=" << nBytesTotal << ", level=" << m_dir_level);
-
-      m_dir_state->set_usage(m_dir_usage_stack.back());
-
-      m_dir_state = 0;
-   }
-
-   void cd_down(const std::string& dir_name)
-   {
-      ++m_dir_level;
-
-      if (m_dir_level <= m_max_dir_level_for_stat_collection)
-      {
-         m_dir_usage_stack.push_back(0);
-         m_dir_state = m_dir_state->find_dir(dir_name, true);
-      }
-
-      m_dir_names_stack.push_back(dir_name);
-      m_current_path.append(dir_name);
-      m_current_path.append("/");
-   }
-
-   void cd_up()
-   {
-      if (m_dir_level <= m_max_dir_level_for_stat_collection)
-      {
-         long long tail = m_dir_usage_stack.back();
-         m_dir_usage_stack.pop_back();
-
-         TRACE_PURGE("FPurgeState::cd_up reporting for '" << m_current_path << "', usage=" << tail << ", level=" << m_dir_level);
-
-         m_dir_state->set_usage(tail);
-         m_dir_state = m_dir_state->get_parent();
-
-         m_dir_usage_stack.back() += tail;
-      }
-
-      // remove trailing / and last dir but keep the new trailing / in place.
-      m_current_path.erase(m_current_path.find_last_of('/', m_current_path.size() - 2) + 1);
-      m_dir_names_stack.pop_back();
-
-      --m_dir_level;
-   }
-
-   // ------------------------------------------------------------------------
-   // ------------------------------------------------------------------------
-
-   FPurgeState(long long iNBytesReq, XrdOss &oss) :
-      nBytesReq(iNBytesReq), nBytesAccum(0), nBytesTotal(0), tMinTimeStamp(0), tMinUVKeepTimeStamp(0),
-      // m_oss(oss),
-      m_oss_at(oss),
-      m_dir_state(0), m_dir_level(0),
-      m_max_dir_level_for_stat_collection(Cache::Conf().m_dirStatsStoreDepth),
-      m_info_ext(XrdPfc::Info::s_infoExtension),
-      m_info_ext_len(strlen(XrdPfc::Info::s_infoExtension)),
-      m_trace(Cache::GetInstance().GetTrace())
-   {
-      m_current_path.reserve(256);
-      m_dir_names_stack.reserve(32);
-      m_dir_usage_stack.reserve(m_max_dir_level_for_stat_collection + 1);
-  }
-
-   // ------------------------------------------------------------------------
-
-   void      setMinTime(time_t min_time) { tMinTimeStamp = min_time; }
-   time_t    getMinTime()          const { return tMinTimeStamp; }
-   void      setUVKeepMinTime(time_t min_time) { tMinUVKeepTimeStamp = min_time; }
-   long long getNBytesTotal()      const { return nBytesTotal; }
-
-   void MoveListEntriesToMap()
-   {
-      for (list_i i = m_flist.begin(); i != m_flist.end(); ++i)
-      {
-         m_fmap.insert(std::make_pair(i->time, *i));
-      }
-      m_flist.clear();
-   }
-
-   /*
-   void UnlinkInfoAndData(const char *fname, long long nbytes, XrdOssDF *iOssDF)
-   {
-      fname[fname_len - m_info_ext_len] = 0;
-      if (nbytes > 0)
-      {
-         if ( ! Cache.GetInstance().IsFileActiveOrPurgeProtected(dataPath))
-         {
-            m_n_purged++;
-            m_bytes_purged += nbytes;
-         } else 
-         {
-            m_n_purge_protected++;
-            m_bytes_purge_protected += nbytes;
-            m_dir_state->add_usage_purged(nbytes);
-            // XXXX should also tweak other stuff?
-            fname[fname_len - m_info_ext_len] = '.';
-            return;
-         }
-      }
-      m_oss_at.Unlink(*iOssDF, fname);
-      fname[fname_len - m_info_ext_len] = '.';
-      m_oss_at.Unlink(*iOssDF, fname);
-   }
-   */
-
-   void CheckFile(const char *fname, Info &info, struct stat &fstat /*, XrdOssDF *iOssDF*/)
-   {
-      static const char *trc_pfx   = "FPurgeState::CheckFile ";
-
-      long long nbytes = info.GetNDownloadedBytes();
-      time_t    atime;
-      if ( ! info.GetLatestDetachTime(atime))
-      {
-         // cinfo file does not contain any known accesses, use fstat.mtime instead.
-         TRACE(Debug, trc_pfx << "could not get access time for " << m_current_path << fname << ", using mtime from stat instead.");
-         atime = fstat.st_mtime;
-      }
-      // TRACE(Dump, trc_pfx << "checking " << fname << " accessTime  " << atime);
-
-      nBytesTotal += nbytes;
-
-      m_dir_usage_stack.back() += nbytes;
-
-      // XXXX Should remove aged-out files here ... but I have trouble getting
-      // the DirState and purge report set up consistently.
-      // Need some serious code reorganization here.
-      // Biggest problem is maintaining overall state a traversal state consistently.
-      // Sigh.
-
-      // In first two cases we lie about FS time (set to 0) to get them all removed early.
-      // The age-based purge atime would also be good as there should be nothing
-      // before that time in the map anyway.
-      // But we use 0 as a test in purge loop to make sure we continue even if enough
-      // disk-space has been freed.
-
-      if (tMinTimeStamp > 0 && atime < tMinTimeStamp)
-      {
-         m_flist.push_back(FS(m_current_path, fname, nbytes, 0, m_dir_state));
-         nBytesAccum += nbytes;
-      }
-      else if (tMinUVKeepTimeStamp > 0 &&
-               Cache::Conf().does_cschk_have_missing_bits(info.GetCkSumState()) &&
-               info.GetNoCkSumTimeForUVKeep() < tMinUVKeepTimeStamp)
-      {
-         m_flist.push_back(FS(m_current_path, fname, nbytes, 0, m_dir_state));
-         nBytesAccum += nbytes;
-      }
-      else if (nBytesAccum < nBytesReq || ( ! m_fmap.empty() && atime < m_fmap.rbegin()->first))
-      {
-         m_fmap.insert(std::make_pair(atime, FS(m_current_path, fname, nbytes, atime, m_dir_state)));
-         nBytesAccum += nbytes;
-
-         // remove newest files from map if necessary
-         while ( ! m_fmap.empty() && nBytesAccum - m_fmap.rbegin()->second.nBytes >= nBytesReq)
-         {
-            nBytesAccum -= m_fmap.rbegin()->second.nBytes;
-            m_fmap.erase(--(m_fmap.rbegin().base()));
-         }
-      }
-   }
-
-   void TraverseNamespace(XrdOssDF *iOssDF)
-   {
-      static const char *trc_pfx = "FPurgeState::TraverseNamespace ";
-
-      char          fname[256];
-      struct stat   fstat;
-      XrdOucEnv     env;
-
-      TRACE_PURGE("Starting to read dir [" << m_current_path << "], iOssDF->getFD()=" << iOssDF->getFD() << ".");
-
-      iOssDF->StatRet(&fstat);
-
-      while (true)
-      {
-         int rc = iOssDF->Readdir(fname, 256);
-
-         if (rc == -ENOENT) {
-            TRACE_PURGE("  Skipping ENOENT dir entry [" << fname << "].");
-            continue;
-         }
-         if (rc != XrdOssOK) {
-            TRACE(Error, trc_pfx << "Readdir error at " << m_current_path  << ", err " << XrdSysE2T(-rc) << ".");
-            break;
-         }
-
-         TRACE_PURGE("  Readdir [" << fname << "]");
-
-         if (fname[0] == 0) {
-            TRACE_PURGE("  Finished reading dir [" << m_current_path << "]. Break loop.");
-            break;
-         }
-         if (fname[0] == '.' && (fname[1] == 0 || (fname[1] == '.' && fname[2] == 0))) {
-            TRACE_PURGE("  Skipping here or parent dir [" << fname << "]. Continue loop.");
-            continue;
-         }
-
-         size_t    fname_len = strlen(fname);
-         XrdOssDF *dfh       = 0;
-
-         if (S_ISDIR(fstat.st_mode))
-         {
-            if (m_oss_at.Opendir(*iOssDF, fname, env, dfh) == XrdOssOK)
-            {
-               cd_down(fname);           TRACE_PURGE("  cd_down -> [" << m_current_path << "].");
-               TraverseNamespace(dfh);
-               cd_up();                  TRACE_PURGE("  cd_up   -> [" << m_current_path << "].");
-            }
-            else
-               TRACE(Warning, trc_pfx << "could not opendir [" << m_current_path << fname << "], " << XrdSysE2T(errno));
-         }
-         else if (fname_len > m_info_ext_len && strncmp(&fname[fname_len - m_info_ext_len], m_info_ext, m_info_ext_len) == 0)
-         {
-            // Check if the file is currently opened / purge-protected is done before unlinking of the file.
-
-            Info cinfo(m_trace);
-
-            if (m_oss_at.OpenRO(*iOssDF, fname, env, dfh) == XrdOssOK && cinfo.Read(dfh, m_current_path.c_str(), fname))
-            {
-               CheckFile(fname, cinfo, fstat);
-            }
-            else
-            {
-               TRACE(Warning, trc_pfx << "can't open or read " << m_current_path << fname << ", err " << XrdSysE2T(errno) << "; purging.");
-               m_oss_at.Unlink(*iOssDF, fname);
-               fname[fname_len - m_info_ext_len] = 0;
-               m_oss_at.Unlink(*iOssDF, fname);
-            }
-         }
-         else // XXXX devel debug only, to be removed
-         {
-            TRACE_PURGE("  Ignoring [" << fname << "], not a dir or cinfo.");
-         }
-
-         delete dfh;
-      }
-   }
-};
 
 const char *FPurgeState::m_traceID = "Purge";
 
@@ -892,7 +437,7 @@ void Cache::Purge()
          size_t      info_ext_len  =  strlen(Info::s_infoExtension);
          int         protected_cnt = 0;
          long long   protected_sum = 0;
-         for (FPurgeState::map_i it = purgeState.m_fmap.begin(); it != purgeState.m_fmap.end(); ++it)
+         for (FPurgeState::map_i it = purgeState.refMap().begin(); it != purgeState.refMap().end(); ++it)
          {
             // Finish when enough space has been freed but not while age-based purging is in progress.
             // Those files are marked with time-stamp = 0.
